--- lib-multisrc/latmanga/assets/i18n/messages_en.properties
@@ -1,4 +1,0 @@
-sort_by_filter_title=Sort by
-sort_by_popular=Popular
-sort_by_latest=Latest
-chapter_prefix=Chapter
