--- lib-multisrc/latmanga/assets/i18n/messages_es.properties
@@ -1,4 +1,0 @@
-sort_by_filter_title=Ordenar por
-sort_by_popular=Populares
-sort_by_latest=Recientes
-chapter_prefix=Capítulo
