--- lib-multisrc/latmanga/build.gradle.kts
@@ -1,9 +1,0 @@
-plugins {
-    id("lib-multisrc")
-}
-
-baseVersionCode = 1
-
-dependencies {
-  api(project(":lib:i18n"))
-}
